requirements.txt
# Core ML & Data Science
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Web App & Visualization
streamlit>=1.20.0
matplotlib>=3.5.0
plotly>=5.0.0
pillow>=9.0.0
opencv-python-headless>=4.5.0

# Medical Imaging
pydicom>=2.3.0
nibabel>=5.0.0

# Advanced Features (Radiomics, Metrics, etc.)
scikit-image>=0.19.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# Reporting
fpdf>=1.7.2
reportlab>=3.6.0
